split up managers for better performance and maintainability
