+++ BloodNight-core/src/main/java/de/eldoria/bloodnight/core/manager/mobmanager/MobManager.java
@@ -1,4 +1,4 @@
-package de.eldoria.bloodnight.core.manager;
+package de.eldoria.bloodnight.core.manager.mobmanager;
 
 import de.eldoria.bloodnight.config.Configuration;
 import de.eldoria.bloodnight.config.worldsettings.WorldSettings;
@@ -7,21 +7,20 @@ import de.eldoria.bloodnight.config.worldsettings.mobsettings.MobSetting;
 import de.eldoria.bloodnight.config.worldsettings.mobsettings.MobSettings;
 import de.eldoria.bloodnight.config.worldsettings.mobsettings.VanillaMobSettings;
 import de.eldoria.bloodnight.core.BloodNight;
+import de.eldoria.bloodnight.core.manager.nightmanager.NightManager;
 import de.eldoria.bloodnight.core.mobfactory.MobFactory;
 import de.eldoria.bloodnight.core.mobfactory.WorldMobFactory;
-import de.eldoria.bloodnight.events.BloodNightEndEvent;
 import de.eldoria.bloodnight.hooks.mythicmobs.MythicMobUtil;
 import de.eldoria.bloodnight.specialmobs.SpecialMob;
 import de.eldoria.bloodnight.specialmobs.SpecialMobUtil;
 import de.eldoria.eldoutilities.entityutils.ProjectileSender;
 import de.eldoria.eldoutilities.entityutils.ProjectileUtil;
+import de.eldoria.eldoutilities.scheduling.DelayedActions;
 import de.eldoria.eldoutilities.threading.IteratingTask;
-import lombok.NonNull;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
 import org.bukkit.block.Beehive;
-import org.bukkit.block.BlockState;
 import org.bukkit.block.data.BlockData;
 import org.bukkit.entity.*;
 import org.bukkit.event.EventHandler;
@@ -37,35 +36,37 @@ import org.bukkit.persistence.PersistentDataType;
 import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 import java.util.logging.Level;
 
-public class MobManager implements Listener, Runnable {
+public class MobManager implements Listener {
     private static final NamespacedKey SPAWNER_SPAWNED = BloodNight.getNamespacedKey("spawnerSpawned");
     private static final NamespacedKey PICKED_UP = BloodNight.getNamespacedKey("pickedUp");
-    private final Map<String, WorldMobs> mobRegistry = new HashMap<>();
     private final NightManager nightManager;
     private final Configuration configuration;
     private final ThreadLocalRandom random = ThreadLocalRandom.current();
     private final Map<String, WorldMobFactory> worldFucktories = new HashMap<>();
-    private final List<Runnable> executeLater = new ArrayList<>();
-    private final List<Runnable> executeNow = new ArrayList<>();
-    private final Queue<Entity> lostEntities = new ArrayDeque<>();
+    private final SpecialMobManager specialMobManager;
+    private final DelayedActions delayedActions;
 
     public MobManager(NightManager nightManager, Configuration configuration) {
         this.nightManager = nightManager;
         this.configuration = configuration;
+        specialMobManager = new SpecialMobManager(nightManager, configuration);
+        specialMobManager.runTaskTimer(BloodNight.getInstance(), 100, 1);
+        BloodNight.getInstance().registerListener(specialMobManager);
+        delayedActions = DelayedActions.start(BloodNight.getInstance());
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onMobSpawn(CreatureSpawnEvent event) {
         if (!nightManager.isBloodNightActive(event.getEntity().getWorld())) return;
 
-        if (event.getSpawnReason() == CreatureSpawnEvent.SpawnReason.SPAWNER
-                && configuration.getGeneralSettings().isSpawnerDropSuppression()) {
+        if (event.getSpawnReason() == CreatureSpawnEvent.SpawnReason.SPAWNER) {
             PersistentDataContainer data = event.getEntity().getPersistentDataContainer();
             data.set(SPAWNER_SPAWNED, PersistentDataType.BYTE, (byte) 1);
-            return;
+            if (configuration.getGeneralSettings().isIgnoreSpawnerMobs()) {
+                return;
+            }
         }
 
         World world = event.getLocation().getWorld();
@@ -83,197 +84,14 @@ public class MobManager implements Listener, Runnable {
 
         mobSettings.getMobByName(mobFactory.get().getMobName());
 
-        executeLater.add(() -> wrapMob(event.getEntity(), mobFactory.get()));
+        delayedActions.schedule(() -> specialMobManager.wrapMob(event.getEntity(), mobFactory.get()), 1);
     }
 
-    public void wrapMob(Entity entity, MobFactory mobFactory) {
-        // This is most likely a mounted special mob. We wont change this.
-        if (!(entity instanceof LivingEntity)) return;
-
-        // I will not try to wrap a special or mythic mob
-        if (SpecialMobUtil.isSpecialMob(entity)) return;
-        if (MythicMobUtil.isMythicMob(entity)) return;
-
-        MobSettings mobSettings = configuration.getWorldSettings(entity.getWorld().getName()).getMobSettings();
-        Optional<MobSetting> mobSetting = mobSettings.getMobByName(mobFactory.getMobName());
-
-        if (!mobSetting.isPresent()) {
-            BloodNight.logger().warning("Tried to create " + mobFactory.getMobName() + " but no settings were present.");
-            return;
-        }
-
-        SpecialMob<?> specialMob = mobFactory.wrap((LivingEntity) entity, mobSettings, mobSetting.get());
-
-        getWorldMobs(entity.getWorld()).put(entity.getUniqueId(), specialMob);
-    }
 
     private WorldMobFactory getWorldMobFactory(World world) {
         return worldFucktories.computeIfAbsent(world.getName(),
                 k -> new WorldMobFactory(configuration.getWorldSettings(world)));
     }
-
-    @Override
-    public void run() {
-        for (World bloodWorld : nightManager.getBloodWorldsSet()) {
-            getWorldMobs(bloodWorld).tick(configuration.getGeneralSettings().getMobTick());
-        }
-
-        // run runnables which should be executed this tick
-        executeNow.forEach(Runnable::run);
-        // clear executed runnables
-        executeNow.clear();
-        // schedule runnables for next tick
-        executeNow.addAll(executeLater);
-        // clear the queue
-        executeLater.clear();
-
-        for (int i = 0; i < Math.min(lostEntities.size(), 10); i++) {
-            Entity poll = lostEntities.poll();
-            if (poll.isValid()) {
-                poll.remove();
-            }
-        }
-    }
-
-    @EventHandler
-    public void onBloodNightEnd(BloodNightEndEvent event) {
-        WorldMobs worldMobs = getWorldMobs(event.getWorld());
-        worldMobs.invokeAll(SpecialMob::onEnd);
-        worldMobs.invokeAll(SpecialMob::remove);
-        worldMobs.clear();
-        IteratingTask<Entity> iteratingTask = new IteratingTask<>(event.getWorld().getEntities(), (e) ->
-        {
-            if (!(e instanceof LivingEntity)) {
-                return false;
-            }
-            if (SpecialMobUtil.isSpecialMob(e)) {
-                lostEntities.add(e);
-                return true;
-            }
-            return false;
-        }, stats -> {
-            BloodNight.logger().config(String.format("Marked %d lost enties for removal in %dms",
-                    stats.getProcessedElements(),
-                    stats.getTime()));
-        });
-
-        iteratingTask.runTaskTimer(BloodNight.getInstance(), 5, 1);
-    }
-
-    /*
-    START OF EVENT REDIRECTING SECTION
-     */
-    @EventHandler
-    public void onEntityTeleport(EntityTeleportEvent event) {
-        getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), v -> v.onTeleport(event));
-    }
-
-    @EventHandler
-    public void onProjectileShoot(ProjectileLaunchEvent event) {
-        ProjectileSender projectileSource = ProjectileUtil.getProjectileSource(event.getEntity());
-        if (projectileSource.isEntity()) {
-            getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(projectileSource.getEntity(), v -> v.onProjectileShoot(event));
-        }
-    }
-
-    @EventHandler
-    public void onProjectileHit(ProjectileHitEvent event) {
-        ProjectileSender projectileSource = ProjectileUtil.getProjectileSource(event.getEntity());
-        if (projectileSource.isEntity()) {
-            getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(projectileSource.getEntity(), m -> m.onProjectileHit(event));
-        }
-    }
-
-    @EventHandler
-    public void onDeath(EntityDeathEvent event) {
-        if (SpecialMobUtil.isSpecialMob(event.getEntity())) {
-            if (SpecialMobUtil.isExtension(event.getEntity())) {
-                Optional<UUID> baseUUID = SpecialMobUtil.getBaseUUID(event.getEntity());
-                if (!baseUUID.isPresent()) {
-                    return;
-                }
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(baseUUID.get(), m -> m.onExtensionDeath(event));
-            } else {
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onDeath(event));
-            }
-        }
-    }
-
-    @EventHandler
-    public void onKill(EntityDeathEvent event) {
-        getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onKill(event));
-    }
-
-    @EventHandler
-    public void onExplosionPrimeEvent(ExplosionPrimeEvent event) {
-        getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onExplosionPrimeEvent(event));
-    }
-
-    @EventHandler
-    public void onExplosionEvent(EntityExplodeEvent event) {
-        getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onExplosionEvent(event));
-    }
-
-    @EventHandler
-    public void onTargetEvent(EntityTargetEvent event) {
-        if (event.getTarget() != null && SpecialMobUtil.isSpecialMob(event.getTarget())) {
-            event.setCancelled(true);
-            return;
-        }
-
-        if (!SpecialMobUtil.isSpecialMob(event.getEntity())) {
-            return;
-        }
-
-        // Block that special mobs target something else than players.
-        // Otherwise they will probably kill each other.
-        if (event.getTarget() != null && event.getTarget().getType() != EntityType.PLAYER) {
-            event.setCancelled(true);
-            return;
-        }
-        if (event.getTarget() == null || event.getTarget() instanceof LivingEntity) {
-            getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onTargetEvent(event));
-        }
-    }
-
-    @EventHandler
-    public void onDamage(EntityDamageEvent event) {
-        if (SpecialMobUtil.isSpecialMob(event.getEntity())) {
-            if (SpecialMobUtil.isExtension(event.getEntity())) {
-                Optional<UUID> baseUUID = SpecialMobUtil.getBaseUUID(event.getEntity());
-                if (!baseUUID.isPresent()) {
-                    return;
-                }
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(baseUUID.get(), m -> m.onExtensionDamage(event));
-            } else {
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onDamage(event));
-            }
-        }
-    }
-
-    @EventHandler
-    public void onDamageByEntity(EntityDamageByEntityEvent event) {
-        if (SpecialMobUtil.isSpecialMob(event.getEntity())) {
-            if (SpecialMobUtil.isExtension(event.getEntity())) {
-                Optional<UUID> baseUUID = SpecialMobUtil.getBaseUUID(event.getEntity());
-                if (!baseUUID.isPresent()) {
-                    return;
-                }
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(baseUUID.get(), m -> m.onDamageByEntity(event));
-            } else {
-                getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getEntity(), m -> m.onDamageByEntity(event));
-            }
-        }
-    }
-
-    @EventHandler
-    public void onHit(EntityDamageByEntityEvent event) {
-        getWorldMobs(event.getEntity().getWorld()).invokeIfPresent(event.getDamager(), m -> m.onHit(event));
-    }
-
-    /*
-    END OF EVENT REDIRECTION SECTION
-     */
 
     /*
      * Handling of vanilla mobs damaging to players during a blood night
@@ -339,20 +157,6 @@ public class MobManager implements Listener, Runnable {
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    public void onItemPickup(EntityPickupItemEvent event) {
-        if (!(event.getEntity() instanceof Monster) && !(event.getEntity() instanceof Boss)) return;
-        if (!configuration.getWorldSettings(event.getEntity().getWorld()).isEnabled()) return;
-        addPickupTag(event.getItem().getItemStack());
-    }
-
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    public void onItemPickup(EntityDropItemEvent event) {
-        if (!(event.getEntity() instanceof Monster) && !(event.getEntity() instanceof Boss)) return;
-        if (!configuration.getWorldSettings(event.getEntity().getWorld()).isEnabled()) return;
-        removePickupTag(event.getItemDrop().getItemStack());
-    }
-
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onEntityKill(EntityDeathEvent event) {
         LivingEntity entity = event.getEntity();
         Player player = event.getEntity().getKiller();
@@ -362,6 +166,12 @@ public class MobManager implements Listener, Runnable {
         if (!nightManager.isBloodNightActive(entity.getWorld())) {
             event.getDrops().forEach(this::removePickupTag);
             return;
+        }
+
+        if (configuration.getGeneralSettings().isIgnoreSpawnerMobs()) {
+            if (entity.getPersistentDataContainer().has(SPAWNER_SPAWNED, PersistentDataType.BYTE)) {
+                return;
+            }
         }
 
         WorldSettings worldSettings = configuration.getWorldSettings(entity.getWorld());
@@ -387,7 +197,7 @@ public class MobManager implements Listener, Runnable {
         // Just remove the entity.
         if (player == null) {
             BloodNight.logger().fine("Entity " + entity.getCustomName() + " was not killed by a player.");
-            getWorldMobs(event.getEntity().getWorld()).remove(event.getEntity().getUniqueId());
+            specialMobManager.remove(event.getEntity());
             return;
         }
 
@@ -472,14 +282,12 @@ public class MobManager implements Listener, Runnable {
 
             }
         }
-        executeLater.add(() ->
-                getWorldMobs(event.getLocation().getWorld())
-                        .remove(event.getEntity().getUniqueId()));
+        delayedActions.schedule(() -> specialMobManager.remove(event.getEntity()), 1);
     }
 
     @EventHandler
     private void onChunkLoad(ChunkLoadEvent event) {
-        WorldMobs worldMobs = getWorldMobs(event.getWorld());
+        WorldMobs worldMobs = specialMobManager.getWorldMobs(event.getWorld());
         for (Entity entity : event.getChunk().getEntities()) {
             Optional<SpecialMob<?>> remove = worldMobs.remove(entity.getUniqueId());
             if (SpecialMobUtil.isSpecialMob(entity)) {
@@ -496,7 +304,7 @@ public class MobManager implements Listener, Runnable {
         AtomicInteger entites = new AtomicInteger(0);
 
         // Bugfix for the sin of flying creepers with bees.
-        IteratingTask<BlockState> blockStateIterator = new IteratingTask<>(
+        new IteratingTask<>(
                 Arrays.asList(event.getChunk().getTileEntities()),
                 e -> {
                     if (e instanceof Beehive) {
@@ -532,16 +340,23 @@ public class MobManager implements Listener, Runnable {
                 },
                 s -> {
                     if (hives.get() != 0) {
-                        BloodNight.logger().config("Checked " + hives.get() + " Hive/s with " + entites.get() + " Entities and removed " + s.getProcessedElements() + " lost bees in " + s.getTime() + "ms.");
+                        BloodNight.logger().fine("Checked " + hives.get() + " Hive/s with " + entites.get() + " Entities and removed " + s.getProcessedElements() + " lost bees in " + s.getTime() + "ms.");
                     }
-                });
-
-        blockStateIterator.runTaskTimer(BloodNight.getInstance(), 0, 1);
+                }).runTaskTimer(BloodNight.getInstance(), 0, 1);
     }
 
-    @NonNull
-    private WorldMobs getWorldMobs(World world) {
-        return mobRegistry.computeIfAbsent(world.getName(), k -> new WorldMobs());
+    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    public void onItemPickup(EntityPickupItemEvent event) {
+        if (!(event.getEntity() instanceof Monster) && !(event.getEntity() instanceof Boss)) return;
+        if (!configuration.getWorldSettings(event.getEntity().getWorld()).isEnabled()) return;
+        addPickupTag(event.getItem().getItemStack());
+    }
+
+    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    public void onItemPickup(EntityDropItemEvent event) {
+        if (!(event.getEntity() instanceof Monster) && !(event.getEntity() instanceof Boss)) return;
+        if (!configuration.getWorldSettings(event.getEntity().getWorld()).isEnabled()) return;
+        removePickupTag(event.getItemDrop().getItemStack());
     }
 
     private void addPickupTag(ItemStack itemStack) {
@@ -572,70 +387,7 @@ public class MobManager implements Listener, Runnable {
         return false;
     }
 
-    private static class WorldMobs {
-        private final Map<UUID, SpecialMob<?>> mobs = new HashMap<>();
-        private final Queue<SpecialMob<?>> tickQueue = new ArrayDeque<>();
-
-        private double entityTick = 0;
-
-        public void invokeIfPresent(Entity entity, Consumer<SpecialMob<?>> invoke) {
-            invokeIfPresent(entity.getUniqueId(), invoke);
-        }
-
-        public void invokeIfPresent(UUID uuid, Consumer<SpecialMob<?>> invoke) {
-            SpecialMob<?> specialMob = mobs.get(uuid);
-            if (specialMob != null) {
-                invoke.accept(specialMob);
-            }
-        }
-
-        public void invokeAll(Consumer<SpecialMob<?>> invoke) {
-            mobs.values().forEach(invoke);
-        }
-
-        public void tick(int tickDelay) {
-            if (tickQueue.isEmpty()) return;
-            entityTick += tickQueue.size() / (double) tickDelay;
-            while (entityTick > 0) {
-                if (tickQueue.isEmpty()) return;
-                SpecialMob<?> poll = tickQueue.poll();
-                if (!poll.getBaseEntity().isValid()) {
-                    remove(poll.getBaseEntity().getUniqueId());
-                    poll.remove();
-                } else {
-                    poll.tick();
-                    tickQueue.add(poll);
-                }
-                entityTick--;
-            }
-        }
-
-        public boolean isEmpty() {
-            return mobs.isEmpty();
-        }
-
-        public void put(UUID key, SpecialMob<?> value) {
-            mobs.put(key, value);
-            tickQueue.add(value);
-        }
-
-        /**
-         * Attemts to remove an entity from world mobs and the world.
-         *
-         * @param key uid of entity
-         * @return special mob if present.
-         */
-        public Optional<SpecialMob<?>> remove(UUID key) {
-            if (!mobs.containsKey(key)) return Optional.empty();
-            SpecialMob<?> removed = mobs.remove(key);
-            tickQueue.remove(removed);
-            removed.remove();
-            return Optional.of(removed);
-        }
-
-        public void clear() {
-            mobs.clear();
-            tickQueue.clear();
-        }
+    public SpecialMobManager getSpecialMobManager() {
+        return specialMobManager;
     }
 }
