+++ BloodNight-core/src/main/java/de/eldoria/bloodnight/core/manager/nightmanager/util/BloodNightData.java
@@ -1,4 +1,4 @@
-package de.eldoria.bloodnight.core.manager.nightmanager;
+package de.eldoria.bloodnight.core.manager.nightmanager.util;
 
 import de.eldoria.bloodnight.config.worldsettings.sound.SoundSettings;
 import de.eldoria.bloodnight.core.BloodNight;
