+++ BloodNight-core/src/main/java/de/eldoria/bloodnight/core/manager/nightmanager/util/ConsistencyCache.java
@@ -1,4 +1,4 @@
-package de.eldoria.bloodnight.core.manager.nightmanager;
+package de.eldoria.bloodnight.core.manager.nightmanager.util;
 
 import lombok.Getter;
 import org.bukkit.OfflinePlayer;
